feat: clean up form
Also cleaned up broken form tests & refactored form modules under home page directory
