+++ client/ui/src/components/index.js
@@ -1,6 +1,5 @@
 import Header from "./Header";
 import Footer from "./Footer";
-import Form from "./Form";
 import Note from "./Note";
 
-export { Header, Footer, Form, Note };
+export { Header, Footer, Note };
