+++ client/ui/src/pages/home/Form/index.js
@@ -1,0 +1,1 @@
+export { default } from "pages/home/Form/Form";
