+++ client/ui/src/pages/home/Form/validateField.js
@@ -1,0 +1,8 @@
+const validateField = (condition, error) => {
+  if (condition) {
+    return error;
+  }
+  return undefined;
+};
+
+export default validateField;
