+++ client/ui/src/pages/home/Form/validateForm.js
@@ -1,0 +1,15 @@
+import validateField from "pages/home/Form/validateField";
+
+const validateForm = (fields) => {
+  let errors = [];
+
+  fields.map( (field) => {
+    const { condition, error } = field;
+    const fieldError = validateField(condition, error);
+    fieldError && errors.push(fieldError)
+  });
+
+  return errors;
+};
+
+export default validateForm;
