+++ client/ui/src/utils/index.js
@@ -1,7 +1,5 @@
-import { validateField } from "./validateField";
-import { validateForm } from "./validateForm";
 import { cleanError } from "./cleanError"
 import { todayDate } from "./todayDate";
 import { api } from "./api"
 
-export { validateField, cleanError, todayDate, validateForm, api };
+export {  cleanError, todayDate, api };
